refactor: styles 상수 분리, 타입 분리
