+++ src/components/icon/Icon.tsx
@@ -1,15 +1,15 @@
 import React, { HTMLAttributes, SVGAttributes } from 'react'
 import { useTheme } from 'styled-components'
 
-import { COLOR } from '@/styles/theme'
 import { ICON_COMPONENT } from './constant'
+import { ColorKey } from '@/styles/type'
 
 interface IconProps
   extends Pick<SVGAttributes<SVGAElement>, 'width' | 'height'>,
     Pick<HTMLAttributes<HTMLDivElement>, 'className' | 'onClick'> {
   type: keyof typeof ICON_COMPONENT
-  stroke?: keyof typeof COLOR
-  fill?: keyof typeof COLOR
+  stroke?: ColorKey
+  fill?: ColorKey
 }
 const Icon = React.forwardRef<HTMLDivElement, IconProps>(
   ({ type, width = 18, height, stroke = 'black', fill, ...restProps }, ref) => {
