--- src/styles/constant.ts
@@ -1,5 +1,0 @@
-export const Z_INDEX = {
-  nav: 100,
-  aside: 50,
-  toast: 200,
-}
