+++ src/styles/constants/color.ts
@@ -1,0 +1,28 @@
+const COLOR = {
+  white: '#ffffff',
+  black: '#000000',
+  warning: '#f66570',
+
+  grey50: '#f9fafb',
+  grey100: '#f2f4f6',
+  grey200: '#e5e8eb',
+  grey300: '#d1d6db',
+  grey400: '#b0b8c1',
+  grey500: '#8b95a1',
+  grey600: '#6b7684',
+  grey700: '#4e5968',
+  grey800: '#333d4b',
+  grey900: '#191f28',
+
+  primary50: '#acbbff',
+  primary100: '#98a7fc',
+  primary200: '#8493e8',
+  primary300: '#707fd4',
+  primary400: '#5c6bc0',
+  primary500: '#4857ac',
+  primary600: '#344398',
+  primary700: '#202f84',
+  primary800: '#0c1b70',
+} as const
+
+export default COLOR
