+++ src/styles/constants/font.ts
@@ -1,35 +1,4 @@
-import { DefaultTheme } from 'styled-components'
-
-import { DEVICE_SIZE } from '@/constants/layout'
-
-export const COLOR = {
-  white: '#ffffff',
-  black: '#000000',
-  warning: '#f66570',
-
-  grey50: '#f9fafb',
-  grey100: '#f2f4f6',
-  grey200: '#e5e8eb',
-  grey300: '#d1d6db',
-  grey400: '#b0b8c1',
-  grey500: '#8b95a1',
-  grey600: '#6b7684',
-  grey700: '#4e5968',
-  grey800: '#333d4b',
-  grey900: '#191f28',
-
-  primary50: '#acbbff',
-  primary100: '#98a7fc',
-  primary200: '#8493e8',
-  primary300: '#707fd4',
-  primary400: '#5c6bc0',
-  primary500: '#4857ac',
-  primary600: '#344398',
-  primary700: '#202f84',
-  primary800: '#0c1b70',
-} as const
-
-export const FONT = {
+const FONT = {
   header_1: {
     fontSize: '36px',
     lineHeight: '1.3',
@@ -87,13 +56,4 @@ export const FONT = {
   },
 } as const
 
-export const MEDIA = {
-  mobile: `@media screen and (max-width: ${DEVICE_SIZE.tablet}px)`,
-  tablet: `@media screen and (max-width: ${DEVICE_SIZE.pc}px)`,
-} as const
-
-export const THEME: DefaultTheme = {
-  color: COLOR,
-  font: FONT,
-  media: MEDIA,
-}
+export default FONT
