+++ src/styles/constants/theme.ts
@@ -1,0 +1,11 @@
+import { DefaultTheme } from 'styled-components'
+
+import COLOR from './color'
+import FONT from './font'
+import MEDIA from './media'
+
+export const THEME: DefaultTheme = {
+  color: COLOR,
+  font: FONT,
+  media: MEDIA,
+}
