+++ src/styles/constants/zIndex.ts
@@ -1,0 +1,7 @@
+const Z_INDEX = {
+  nav: 100,
+  aside: 50,
+  toast: 200,
+} as const
+
+export default Z_INDEX
