+++ src/styles/styled.d.ts
@@ -1,6 +1,9 @@
 import 'styled-components'
-import { COLOR, FONT, MEDIA } from './theme'
 import {} from 'styled-components/cssprop'
+
+import COLOR from './constants/color'
+import FONT from './constants/font'
+import MEDIA from './constants/media'
 
 declare module 'styled-components' {
   export interface DefaultTheme {
