+++ src/styles/type.ts
@@ -1,0 +1,8 @@
+import COLOR from './constants/color'
+import FONT from './constants/font'
+
+export type Color = typeof COLOR
+export type ColorKey = keyof Color
+
+export type Font = typeof FONT
+export type FontKey = keyof Font
